FAGSYSTEM-352387: Litt refaktorisering + tester.
